resolved Omri's suggestions
